requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
python-dotenv
httpx
pydantic
